added requested changes and fixed url in index.html
