+++ index.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Nice Gadgets</title>
-    <link rel="icon" type="image/png" href="/icons/favicon.png">
+    <link rel="icon" type="image/png" href="./icons/favicon.png">
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/index.tsx"></script>
+    <script type="module" src="./src/index.tsx"></script>
   </body>
 </html>
